pyproject.toml
[project]
name = "addis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "google-cloud-aiplatform[adk,agent-engines]>=1.71.1",
    "mcp[cli]>=1.25.0",
    "psycopg2-binary",
    "python-dotenv",
    "streamlit>=1.52.2",
    "vertexai>=1.71.1",
]
